+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -157,7 +157,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added final image histogram debug view (both luminance and RGB).
 - Added an example texture and rotation to the Cloud Layer volume override.
 - Added an option to extend the camera culling for skinned mesh animation in ray tracing effects (1258547).
-- Added decal layer mask for decal projectors and material. Materials will receive a decal when both decal layer mask matches. (intersection is not null).
+- Added decal layer system similar to light layer. Mesh will receive a decal when both decal layer mask matches.
 
 ### Fixed
 - Fix when rescale probe all direction below zero (1219246)